requirements.txt
Flask
pybadges
gunicorn
psycopg2-binary